Add files via upload
This isn't a completed version of the index file for the Kimchi Soxchain website but just a progress update for you to look at and provided feedback. I still need to add new images, links, and more caption to the code.  Thank you in advance for your time and feedback.